refactor: 리팩토링
- 배치 실행시 디비 삭제안하도록
- productIds -> playerIds
- hiring -> scout
