+++ soccer-api/src/main/kotlin/com/example/soccerapi/team/api/request/ScoutRequest.kt
@@ -1,5 +1,5 @@
 package com.example.soccerapi.team.api.request
 
 data class ScoutRequest(
-    val productIds: List<Long>,
+    val playerIds: List<Long>,
 )
